Refactor: extract shared types, DRY mergeProperties, fix test races, split manifest
- Extract hubspotObjectRequest, hubspotObjectResponse, mergeProperties into types.go
- Refactor create_contact, create_deal, create_ticket to use shared mergeProperties
- Fix test race conditions: protect shared `calls` slice with sync.Mutex in
  add_note_test.go and create_deal_test.go
- Split large Manifest() method (~230 lines) into manifest.go to reduce hubspot.go size
