+++ connectors/hubspot/create_contact.go
@@ -34,34 +34,20 @@ func (p *createContactParams) validate() error {
 // toAPIProperties merges explicit fields and the additional properties map
 // into the flat properties object that HubSpot expects.
 func (p *createContactParams) toAPIProperties() map[string]string {
-	props := make(map[string]string)
-	// Copy additional properties first so explicit fields take precedence.
-	for k, v := range p.Properties {
-		props[k] = v
-	}
-	props["email"] = p.Email
+	overrides := map[string]string{"email": p.Email}
 	if p.FirstName != "" {
-		props["firstname"] = p.FirstName
+		overrides["firstname"] = p.FirstName
 	}
 	if p.LastName != "" {
-		props["lastname"] = p.LastName
+		overrides["lastname"] = p.LastName
 	}
 	if p.Phone != "" {
-		props["phone"] = p.Phone
+		overrides["phone"] = p.Phone
 	}
 	if p.Company != "" {
-		props["company"] = p.Company
+		overrides["company"] = p.Company
 	}
-	return props
-}
-
-type hubspotObjectRequest struct {
-	Properties map[string]string `json:"properties"`
-}
-
-type hubspotObjectResponse struct {
-	ID         string            `json:"id"`
-	Properties map[string]string `json:"properties"`
+	return mergeProperties(p.Properties, overrides)
 }
 
 func (a *createContactAction) Execute(ctx context.Context, req connectors.ActionRequest) (*connectors.ActionResult, error) {
